[#193] feat : 초대받은 내역 서버렌더링
